v2.0.2 -> v2.0.3

+++ sub_apps/observation.py
@@ -1,7 +1,8 @@
 import asyncio
 from datetime import datetime
 import logging
-from typing import List
+from typing import (List,
+                    Tuple)
 
 from aiogram.utils.exceptions import (BotBlocked,
                                       ChatNotFound,
@@ -16,8 +17,7 @@ from models.localization import Localization
 from utils.intra_api import (IntraAPI,
                              NotFoundIntraError,
                              UnknownIntraError)
-from utils.cache import (Cache,
-                         cache)
+from utils.cache import Cache
 
 
 class Observation:
@@ -26,14 +26,13 @@ class Observation:
         self._local = local
         self._logger = logging.getLogger('Observation')
 
-    @cache(ttl=3600)
-    async def _get_observables(self, limit: int, offset: int) -> List[list]:
-        observables = await db.select([Peer.login, db.func.array_agg(UserPeer.user_id)]).select_from(
+    @staticmethod
+    async def _get_observables(limit: int, offset: int) -> Tuple[str, List[int]]:
+        return await db.select([Peer.login, db.func.array_agg(UserPeer.user_id)]).select_from(
             Peer.join(UserPeer)).where(UserPeer.relationship == Relationship.observable).limit(limit).offset(
             offset).group_by(Peer.id).gino.all()
-        return [list(observable) for observable in observables]
 
-    async def _send_notifications(self, observables: List[list]):
+    async def _observation_process(self, observables: Tuple[str, List[int]]):
         from misc import bot
 
         for login, user_ids in observables:
@@ -41,7 +40,7 @@ class Observation:
             try:
                 peer = await self._intra.get_peer(login=login)
             except (NotFoundIntraError, UnknownIntraError) as e:
-                self._logger.error('Response exception=%s %s, continue next', login, e)
+                self._logger.error('Error response=%s %s, continue next', login, e)
                 continue
             if peer:
                 location = await Cache().get(key=f'Location:{login}')
@@ -53,25 +52,30 @@ class Observation:
                 truths = (current_location != 'not on campus', location is not None, location != current_location)
                 if all(truths):
                     for user_id in user_ids:
-                        *_, user = await User.get_user_data(user_id=user_id)
-                        if isinstance(user, dict):
-                            user = User.from_dict(data=user)
-                        try:
-                            text = self._local.in_campus.get(user.language, login=login,
-                                                             current_location=current_location)
-                            await bot.send_message(chat_id=user_id, text=text)
-                            self._logger.info('Send notification=%s %s',
-                                              login, user.username or user.id)
-                            await asyncio.sleep(0.1)
+                        self._logger.info('Trying to send notification=%s %s', login, user_id)
+                        user_data = await User.get_user_data(user_id=user_id)
+                        if user_data:
+                            *_, user = user_data
+                            if isinstance(user, dict):
+                                user = User.from_dict(data=user)
+                            try:
+                                text = self._local.in_campus.get(user.language, login=login,
+                                                                 current_location=current_location)
+                                await bot.send_message(chat_id=user_id, text=text)
+                                self._logger.info('Send notification=%s %s',
+                                                  login, user.username or user.id)
+                                await asyncio.sleep(0.1)
 
-                        except (BotBlocked, UserDeactivated) as e:
-                            await user.delete()
-                            self._logger.error('BotBlocked or UserDeactivated user=%s %s, user deleted',
-                                               user.username or user.id, e)
+                            except (BotBlocked, UserDeactivated) as e:
+                                await user.delete()
+                                self._logger.error('BotBlocked or UserDeactivated user=%s %s, user deleted',
+                                                   user.username or user.id, e)
 
-                        except ChatNotFound as e:
-                            self._logger.error('ChatNotFound user=%s %s',
-                                               user.username or user.id, e)
+                            except ChatNotFound as e:
+                                self._logger.error('ChatNotFound user=%s %s',
+                                                   user.username or user.id, e)
+                        else:
+                            self._logger.error('User not found user=%s', user_id)
                 self._logger.info('Complete observation process for peer=%s', login)
 
     async def observation(self):
@@ -81,7 +85,10 @@ class Observation:
             offset = 0
             observables = await self._get_observables(limit=100, offset=offset)
             while observables:
-                await self._send_notifications(observables=observables)
+                try:
+                    await self._observation_process(observables=observables)
+                except Exception as e:
+                    self._logger.error('Unknown observation error %s', e)
                 offset += 100
                 observables = await self._get_observables(limit=100, offset=offset)
             self._logger.info('Complete observation')
